Update README.md
MTN (Mobile Telephone Network) is a telecommunication industry in Nigeria that is in the business of providing voice services via 2G, 3G and 4G networks, including prepaid and postpaid airtime (on various price plans), message service, international roaming, electronic voucher distribution service and more.
The dashboard shows the data type,the amount and the monthly subscriber Technology data available with MTN between April 2017 and February 2022. It was discovered that there was up 124.05% subscription prior to COVID'19 and the increase comes with the festive period and the early months of the year, after which the percentage of the subscription reduces midyear,so it can be concluded here that perhaps, a reasonable number of persons comes back to Nigeria to celebrate the season which leads to the rise and fall in subscription.

After Covid'19, it was realized that there has never been any such increase in the subscription again as the highest recorded is currently 108.94%. This could be as a result of the series of lock down experienced and the processes involved before one can travel from one country to the other and in addition, this could be as a result of terrorist attack that has taken so many lives in Nigeria and the number of persons who has been pushed to poverty and probably have moved to a remote area where her is no network to assess the internet.

It is also discovered that customers tend to subscribe more to cheaper data plans (for example 2 day plan) with higher turnover than expensive data plans. Please see the attached for more.